Quick Open: Update provider results prior to calling setState in handleProvidersChange
Summary:
Move the setState call in QuickSelectionComponent.handleProvidersChange() to happen after
calling _updateResults(). Previously, when a provider is removed, the setState call would
immediately trigger render, which in turn will attempt to retrieve providers from a stale list.
This caused errors when disabling a provider package, because render tries to access a
non-existent provider, and the invariant in SearchResultManager._getProviderByName will throw.
Calling _updateResults first ensures that render is not triggered before the provider list is
updated.

Reviewed By: jgebhardt

Differential Revision: D3292257

fbshipit-source-id: d3480c20d2d6e7605f45dee64c8a60b482c85ed2