Use type-coverage package for Hack type coverage
Summary:
In practice the percentage seems to be a little bit on the low side so I think
the way it's computed merits a bit more investigation. The test show that it's
technically correct but it doesn't quite feel right.

Also note that I'm leaving the busy-signal and track timing in, even though in
practice it seems to be fast enough that busy-signal isn't really necessary. The
usage analytics from the timing call will stop being useful since it will now be
called whether people are paying attention to it or not, just to generate the
status bar widget. But the timing analytics will continue to be useful.

Reviewed By: peterhal

Differential Revision: D3298762

fbshipit-source-id: 21f06f05314ac5ec9fad7c5e140e06edee01b6b6