getProperties api for the WatchExpressionStore
Summary:
This adds a method on the WatchExpressionStore to fetch properties for a given objectId string.  Most of the work of this diff is in generalizing the existing code used for evaluation since much of the underlying process between `getProperties` and `evaluate` is the same.  A notable difference is that `EvaluationResponses` Observable returned from the store can be updated with new values when the `triggerReevaluation` method is called, but there is no way to "re-evaluate" properties.  This isn't needed, because the consumer (e.g. the lazy nested value component) should instead:

1. re-evaluate their expression
2. get a new `objectId` from the returned `EvaluationResult`
3. unsubscribe from the previous `getProperties` observable
4. call `getProperties` on the store again with the new `objectId` from step 2

Reviewed By: yinghuitan

Differential Revision: D3290827

fbshipit-source-id: a297d939d0972c03cf117009b819be318d8753b8