Display native notification when a breakpoint is hit
Summary:
Here is a common frustration when debugging:

1. Set breakpoint and attach debugger
2. Forget that you've done this
3. Refresh some page in your sandbox only to find that it is hanging and won't load
4. Get confused/mad/sad until you realize it's because the debugger paused your request

jamesgpearce just included the ability to display native notifications from atom, so the idea of this diff is to
display a native notification when breakpoints are triggered.  That way when you refresh the page, you'll
have a clue as to why it's not loading.

Thoughts?

Reviewed By: jamesgpearce

Differential Revision: D3286224

fbshipit-source-id: 8decef047abfeb39ba872504b50bdaf7ba3ffa33