Make header bar participate in stacking context
Summary:
This commit fixes a z-index issue on [http://nuclide.io](http://nuclide.io) as viewed on small screens.

Currently, the .headerBarContainer has a z-index of 9999, but when styled as a hamburger menu, it does not meet [the criteria](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Positioning/Understanding_z_index/The_stacking_context) to create a stacking context.

As a result, the z-index doesn't have the expected effect: .promoSection is rendered above the opened hamburger menu:

<img width="396" alt="screen shot 2016-05-15 at 9 06 58 pm" src="https://cloud.githubusercontent.com/assets/6564473/15278065/11e81242-1ae1-11e6-990c-711d1b60c9d0.png">

The simplest fix is to position the element.
Closes https://github.com/facebook/nuclide/pull/498

Reviewed By: JoelMarcey

Differential Revision: D3306566

Pulled By: ssorallen

fbshipit-source-id: 69ad02cecccc8645fda4e989ec961f08a9ca7b87