nuclide-atom-script Atom 1.7 compat and add proper console logging
Summary:
This is effectively a re-write of `nuclide-atom-script`. The main goal was to remove the console-via-a-socket hack, but along the way, the dependency on Node was removed and using application state in Atom 1.7+ was fixed.

The way `console` works now is by first turning off Chromiums verbosity with `--v=-3`, and then using the `write-to-stdout` and `write-to-stderr` ipc channels for user code calls to `console`. A new `console` object is constructed using the `Console` constructor, with pseudo-streams that write to the ipc channels. This preservers all of the semantics of the `console` methods - like `%s` and `console.dir`. `log`, `info` and `dir` go to standard out, and `error`, `warn` and `trace` go to standard error.

Since there's no socket to maintain, there's no need for a node script outside of Atom. Now there's just a small bash script to kick things off.

In Atom 1.7, the user data dir is set to a temp dir when `--test` is used. This meant that things like cookies that existed in a user's normal Atom env, didn't exist in the test script. This is now fixed by explicitly passing `--user-data-dir`.

The only thing that still needs a bit of work here, is figuring out a better way to pass arguments to the test script. Right now it's done through an environment variable. This is fine, except that you lose quoting.

Reviewed By: bolinfest

Differential Revision: D3304450

fbshipit-source-id: f14b1a1eeeebbaa1dac37d1b9e9d72e82bde241d