Add DebuggerSettings to config Threads window visibility
Summary:
A common requirement in Nuclide debugger is supporting different settings for different languages/platforms. For example, threads window is supported in C++ while not php/JS(at least not yet).
This diff added a DebuggerSettings channel between frontend and backend so that we can toggle different UI states.

Ideally we should extend chrome protocol to directly query backend for capability support but I would wait that for new chrome UI and it is always useful to have a channel to push some debugger settings to backend.

Note: I did not toggle SupportThreadsWindow yet, that's for another diff.

Reviewed By: johnislarry

Differential Revision: D3295566

fbshipit-source-id: 4d4f580279586e967d91815cd6c96dff2b4697b9