Fix failing assert on update from <v0.136.0
Summary:
Version >=0.137 have protocol field in their messages, messages <=0.136 don't.
When the client connects to the server it is possible that the server is of an older version (and needs to be resterted)
This diff comments out an assertion that fails on servers from <=0.136 until we can be reasonably sure there are no such left.

Reviewed By: peterhal

Differential Revision: D3311419

fbshipit-source-id: ad1f9385e2453a22abca6f1f09b5d33a0ded8909