Buck build: display loading indicator when fetching rule info
Summary:
Add a status indicator that shows one of two states:

1. a loading spinner, when we're fetching rule info
2. a warning dialog, when no matching rule is found

If we've fetched the rule and it's valid, then show the iOS simulator + RN checkbox as usual.

There's a bit of unwanted state around counting pending Promises. Hopefully I'll convert this into an Observable stream to avoid that eventually.

Reviewed By: ssorallen

Differential Revision: D3286390

fbshipit-source-id: 221933436c62e8897a507adcb084fa3fcb50aac5