Button to disable warnings that Flow is not installed
Summary:
People might not be interested in Flow, may just be code browsing, or making
small changes to existing JS projects that use Flow, and therefore won't want to
install Flow but will want to browse code in peace. Therefore it's useful to
have a setting to disable this popup.

Reviewed By: zertosh

Differential Revision: D3311994

fbshipit-source-id: 66b4030f7c6d4d869b67f6fefcf1aa4321c037fc