Created Nux package with basic View and Tour components.
Summary:
Creation of the basic infrastructure for a New User Experience (NUX) framework for Nuclide.

Additions include:
  - NuxView component: Used to attach a NUX to the DOM and display content. Uses atom tooltips as underlying mechanic for attaching content as a NUX
  - NuxTour component: Used to house a collection of NuxViews and manage progression through a guided NUX tour. Also manages logging and tracking.

A new NUX can be created using a query selector (or selector function), direction for the tooltip to appear, content, whether the content is HTML or not, and display / completion predicates.

Display predicates are used to evaluate whether or not to display the NuxView. If the predicate fails, it will retry up to (configurable) 4 times waiting (also configurable) 500ms between, and if it fails after all of them then an Error will be thrown. Completion predicates are similar, but determine whether or not the NUX has been "completed", e.g. whether a panel has popped open in response to a button click. These are very useful when assembling NUXs inside a NuxTour.

Reviewed By: nmote

Differential Revision: D3308609

fbshipit-source-id: 0fb74bce5856d413fcda926938aed02337bc769e