Clang: apply timeout + thread limit to compilation flags process
Summary:
People have been reporting instances where the compilation flags build is taking up a large amount of resources. I suspect this is mostly due to cold disk cache, but as a background process we should lighten up on resource usage.

Apply a thread limit of 8 and a timeout of 60s.

Reviewed By: zertosh

Differential Revision: D3321258

fbshipit-source-id: a1872ece4aae4ccc529320a05216f740876fb020