Select the first frame with source code during break
Summary:
While showing callstack for debugger break chrome displays the first frame. This maybe fine for JS/php because you always have source code for each frame, but may not be ideal for C++ because the top frames may not have sources.
This diff adds "hasSource" into callstack/threads protocol so that we automatically search the first frame with source code and select it.
Note: this parameter is optional and returns true by default so it won't affect php/JS.

Reviewed By: johnislarry

Differential Revision: D3323729

fbshipit-source-id: ab2a0a102a34782cc88d277f94cdfaadc5f7d912