Emit 'did-change-short-head' to watch changing Mercurial bookmark
Summary:
The `HgService` watches the active bookmark already but emits only the
broad `'did-change-statuses'` when it changes. If the client knows
they are working with a Mercurial repository instead of a Git one (the
implementation from GitHub), they can instead listen for
`'did-change-short-head'` if they want a event for when only the active
bookmark changes.

Reviewed By: mostafaeweda

Differential Revision: D3322764

fbshipit-source-id: 3d4867ffd3b4c16cd9f25b18e5badd8cf8014279