Report error to user when we recieve a malformed debugger response
Summary:
Report an error to the user (in the form of an atom notification) when we have trouble parsing HHVM
messages in the DbgpConnector.  We do this in the DbgpSocket, as well, so this behavior would be consistent.
It would also lead to a better experience, because the user would know more quickly that the debugger had an
issue.

Reviewed By: yinghuitan

Differential Revision: D3325793

fbshipit-source-id: 20e04b2ec7f46ffcef05f58e2512271b27b3929e