Adding representativeName property to the outline and to providers
Summary:
Adds support for specifying a representative name as part of an outline.
The representative name is the most obvious name that makes sense for a line of the outline tree (it may also be missing).

For example, for a line listing a method with arguments, method name would make the most sensible representative name.

Reviewed By: nmote

Differential Revision: D3296406

fbshipit-source-id: c761dcc301a4474a9f0ba5470414199586ae0374