Populate symbols-view with the representative names from the outline
Summary:
In a rather hackish way - inject the items from the outline view to the file-view panel of the symbols-view.

The functionality is hidden behind a setting marked as experimental (and probably will be such for quite a long time).

If either the property is turned off, or the symbols-view package is not enabled, the injection is not performed.

Reviewed By: nmote

Differential Revision: D3296419

fbshipit-source-id: 81fe0832134d863fdb2ae4359148ddeefa6e8506