Buck cleanup: move build logic from store to build system
Summary:
This refactors the build logic into the build system. We shouldn't be keeping track of all this stuff in the store - in particular, we no longer keep track of the build progress and React Native state managers.

Much of the logic was also rewritten to make it more observable-friendly.

Reviewed By: ssorallen

Differential Revision: D3311839

fbshipit-source-id: 156a45f7902227e4126fae89aba42816f747779e