Add a --dev option to the setup script.
Summary:
The goal in introducing this option is for making it so that `atom` always runs
a "stable" version of Nuclide that is installed the normal way whereas `atom --dev`
loads the version of Nuclide you are developing locally. This is helpful so that
you can do all sorts of crazy Nuclide development while always having a reliable,
working Nuclide handy.

Now when `./scripts/dev/setup --dev` is run, it symlinks `~/.atom/dev/packages/nuclide`
instead of `~/.atom/packages/nuclide`.

Reviewed By: zertosh

Differential Revision: D3332329

fbshipit-source-id: 8a8d40488d7c500c143c4d0b3e186e5221f67c84