Update path in TextEditors when renaming a file
Summary: TextEditors need to be updated if a user renames a file from the file-tree. Otherwise, if they have an unsaved content and user hits save after a rename, the file is effectively duplicated. One time with the old name and the changes and one time with the new name and without.

Reviewed By: mostafaeweda

Differential Revision: D3330591

fbshipit-source-id: 0406ebbc1d32becad48dc28aa61acac513505d1f