Add delay prop to LoadingSpinner
Summary:
We often want to show spinners only if we're waiting a substantial amount of time to reduce visual churn.

`<LoadingSpinner delay={200} />` will not render anything until 200ms have passed. I think this is a very common use case for the Loading Spinner, so I chose to build it into this component (rather than an HOC).

Reviewed By: yinghuitan

Differential Revision: D3320967

fbshipit-source-id: ad6dd8d0baee3beb60a2cdaac65a8b5b4a2f0a65