Created NuxManager to react to store changes and manage displaying of NUXes to the user.
Summary:
The NuxManager handles updates from the NuxStore. The goal in the future is to be able to check whether the new NUX passes a Gatekeeper, after which the Manager will add the NUX to a queue.

The queue can then be consumed at appropriate times/moments and NUXes can be displayed when appropriate. For now, the NUX is shown as soon as it is available (while we only have one NUX to show on launch).

Reviewed By: jgebhardt

Differential Revision: D3336957

fbshipit-source-id: a44e5811c5b9c225aee043bc44d473037fda3a61