Remove erroneous comment
Summary:
D3314170 redoes the logic for sending the 'stop' xdebug command.  The logic the comment I removed is referring to happens when we have a 'bottom up' killing of the debugger (i.e. we run out of requests or a script finishes execution, and the debugger decides to exit on its own).  I used to think that we should have a guarantee that any scripts have fully exited by this point once we were able to send the stop command, and therefore there would be no need to await.

Unfortunately, I just removed the logic and tested this by running a script, and it doesn't work.  It actually crashes the server.  Here is the sequence of events:

1. Script finishes, Nuclide receives response for stop command, connection is removed and we hit the checkForEnd function
2. An end event is emitted from the mux and the session is ended (or partially ended since this is a race)
3. The script process exits and calls its callback with its exit code, but the output window observable has been unsubscribed so we throw a fatal error

So it seems to me that we still do need this await, since even if we are given a response for the stop command from hhvm, it doesn't guarantee that the process has fully exited yet.

Reviewed By: jgebhardt

Differential Revision: D3318395

fbshipit-source-id: 4ab5bc9a83c527a31d525efca197abda2ab4223f