Diff View: Fix buffer contents override and state-corruption between closing/re-opening
Summary:
At a prior Atom/Diff-View version, `pane.destroy()` caused unmounting to the hosted react component (now that's no longer the case).
With the child component buffer change subscription not being disposed, the model state was changing while the diff view is closed.
That, and setting the buffer contents to `newContents` on re-opening the diff view caused both:
* 2 change subscriptions are setup and conflicting model's `newContents` state.
* Override the next-to-load buffer with whatever buffer edited outside of the Diff View, triggering the model-change subscription.

This diff makes sure that the child components are unmounted when the parent component is unmounted.
Also, the diff view code should never change the active buffer text contents - the model-pushed contents should only be used for offset and highlight calculation/rendering.

Reviewed By: ssorallen

Differential Revision: D3354670

fbshipit-source-id: 864d00fd3d3c2016f3b10eed02532e83e4f3bf55