Diff View: warning and reset diff state when the active buffer is destroyed
Summary:
Unlike pane item text editors, the Diff View didn't listen to buffer destroy events, indicating the undelying files being removed.
Hence, re-creating the buffer (but not reloading it) on re-render, hence no digest cache was found, leading to a text buffer serialization error.
With this, the Diff View's expected behavior when deleting a file is to warn the user and reset the diff state.

Reviewed By: jamesgpearce

Differential Revision: D3353431

fbshipit-source-id: f3065026c3a92c5c66deafd268818ce68e228e2a