Merge nuclide-ssh-dialog into nuclide-remote-projects
Summary: This fixes an issue where the first time you open the remote connection dialog, the dialog appears to jump (expand then contract). What's happening is that `nuclide-ssh-dialog` is not an Atom package, but it depends on styles. It uses `nuclide-atom-apm` to inject the styles on load. So what you're seeing is an unstyled password input, that then restyles. By merging `nuclide-ssh-dialog` into `nuclide-remote-projects` (a real Atom package), we let Atom load the styles at startup.

Reviewed By: ssorallen

Differential Revision: D3355621

fbshipit-source-id: c0bfe51f1214b6923e11e5852aa0408bc8b7c207