Change default AtomInput width; handle null value
Summary:
Previously, there was a bug whereby the width passed to the AtomInput wouldn't actually be used (unless/until it was changed to a different value). This was fixed in D3356094 but it turns out that things like the connection dialog were relying on the broken behavior.

This updates the component in the following ways:

* Support `null` as a value for the width prop. Previously there was no way to remove an explicit width (it only worked by virtue of never having correctly set it in the first place).
* Change the default width to `null`. The previous default (200) wasn't actually being used so this maintains the old behavior.

Reviewed By: zertosh

Differential Revision: D3357495

fbshipit-source-id: f98348be90f097eaae4d18380484f143d526557b