NuclideSocket: schedule reconnect after startup errors
Summary:
Normally we automatically schedule a reconnect when the transport closes, but there's a corner case here: if the socket errors on startup, the transport never gets connected and we're stuck with a socket in a zombie state.

If we're fortunate, the heartbeat will be restored and the socket gets reconnected anyway. It's definitely possible that this doesn't happen though.

Reviewed By: zertosh

Differential Revision: D3359018

fbshipit-source-id: 79c47b1d56ac676c938a0274fb547ce511455431