nuclide-diagnostics-ui remove keyboard shortcut label and move file filter to right
Summary: This makes the UI more consistent since we don't give keyboard shortcut hints anywhere else (keyboard users know where to find shortcuts). By moving the file filter toggle to the right, changes to the count as a result of clicking on it, won't result in it shifting as a result of the count width change.

Reviewed By: nmote

Differential Revision: D3364003

fbshipit-source-id: 9bf9c84f98700c8cf7847161e69f3ea73fdf5846