Warmup transpile cache before running tests
Summary:
This diff changes the default behavior of the `release_transpile.js` script to not overwrite the files it's transpiling (to overwrite, as you'd need for a release, pass in `--overwrite`). This has the effect of warming up the transpile cache w/o destroying your working files.

Changes to `release_transpile.js`:

* No longer ignores `**/spec/**`. This was an optimization for production builds. But we need those files in the cache now.
* Transpiling happens across worker processes. So much faster! For this:
  * `NodeTranspiler` got a callback to notify when writing to the cache was done.

Reviewed By: ssorallen

Differential Revision: D3362761

fbshipit-source-id: 6c62d9f57d38dfcf05ee9879ba3cd6233ef46e10