Reverted commit D3359018
Summary:
Normally we automatically schedule a reconnect when the transport closes, but there's a corner case here: if the socket errors on startup, the transport never gets connected and we're stuck with a socket in a zombie state.

If we're fortunate, the heartbeat will be restored and the socket gets reconnected anyway. It's definitely possible that this doesn't happen though.

Reviewed By: zertosh

Differential Revision: D3359018

fbshipit-source-id: a2c8644cd4e4255d0871530025fe83449c4ac938