Fix remote integration test hanging
Summary:
It looks like the 100ms delay for the server close response is too small - sometimes the server still gets killed before the response can be sent. Increase it to 1s.

I think this affects only the clipboard test, since it never actually issues any remote commands. When this is the case, I think we get to `stopNuclideServer` while the server is still starting up..

Reviewed By: zertosh

Differential Revision: D3373401

fbshipit-source-id: 95440291fa57ef4eaea0629d443464ce0be1073c