Diff View Speedup: fetch commit file changes only when needed
Summary:
In compare's default case + commit/amend modes: we only need to display the dirty file changes (should be fast, not waiting for any other file changes to load).
In publish view, only the last commit's changes is needed.
In compare view, only the user-selected commit changes should be requested (especially when working on a big stack).

Here, instead of managing 3 types of files changes, this brings the knowledge and update code down to the repository stack, hence, it can optimally decide when to invalidate and what changes to re-query.
This is basically a re-write of the `RepositoryStack` and its API integration with the `DiffViewModel`.

Reviewed By: ssorallen

Differential Revision: D3368957

fbshipit-source-id: 5d8da393414ac65934e9fa168121315f5cea55ed