Don't collapse a directory on failed fetch attempt if contents were fetched in the past
Summary:
Usually file-tree collapses a directory if its contents could not be fetched. This, however, might lead to a poor UX on a flaky network, when a directory was fetched successfully in the past, but a subsequent refetch (due to a collapse/expand, for example) is failing.

To make things worse, the collapse is only preformed after a hugely long timeout of 60 seconds, and for all purposes comes as out of the blue for the user.

This diff adds a `wasFetched` property to the FileTreeNode and uses it to prevent the automatic collapse on failing subsequent refetches.

Reviewed By: hansonw

Differential Revision: D3349623

fbshipit-source-id: 4be312bc88963167b2dfa30799cbb024f70e7448