Tell flow to ignore missing facebook modules
Summary:
This PR tells `flow` not to report errors for the various places where `fb` modules are required but may not be available. I'm assuming these are modules for internal use at facebook since they are not installed with the other dependencies and, as far as I could tell, not mentioned in the documentation.
Closes https://github.com/facebook/nuclide/pull/527

Reviewed By: zertosh

Differential Revision: D3374324

Pulled By: nmote

fbshipit-source-id: a8c10b5edb70c1347c6c6c9abbd6a1768fd1c11a