Clang server: move flag change management into ClangFlagsManager [6/n]
Summary:
After we fetch compilation flags for a C++ file, we maintain a watch on its owning Buck/TARGETS file so that we can update them when the file changes.

The subscription is currently managed by the `ClangServer`, but it should really be part of `ClangFlagsManager`. In the end state, `ClangServer` just needs to receive a raw list of flags.

Reviewed By: peterhal

Differential Revision: D3343914

fbshipit-source-id: e9b57c1f837c72fe03be8c3f4a32f0b6feda8022