Clang server: pass in flags at startup (and only at startup) [8/n]
Summary:
We currently pass along compilation flags with every request, but this is actually not very useful: only the first request actually uses it. Once we've compiled a file with a given list of flags, updating the flags requires a full restart.

Instead, just pass along the flags at startup via the command line.

Reviewed By: peterhal

Differential Revision: D3343985

fbshipit-source-id: 14e479b91a9057a8dc1cdb0d317290610193f92a