Clang server: refactor server management into a helper class [9/n]
Summary:
This creates a `ClangServerManager` class that handles the messy details around managing flags, limiting the number of servers, and keeping memory usage in check.

As a bonus, this is also much more unit testable. I added some tests for some of the tricker cases we handle. (Note that `waitsForPromise` is currently unused, but it will be needed in the next diff).

Reviewed By: peterhal

Differential Revision: D3350761

fbshipit-source-id: e0474d1bb82dd9231fd4006b0d6fcc681ba96d58