Clang server: finish migration to RpcProcess [11/11]
Summary:
Now that all the custom logic has been moved out, we can use tyangliu's reusable RpcProcess abstraction to interface with the Python Clang server. So much red code!

There's only one quirk which sticks out - the fact that we normally have requests totally bounce while the server's compiling. Ideally this should be handled like FlowProcess with an external status indicator, but this is probably the easiest solution for now.

Reviewed By: peterhal

Differential Revision: D3350985

fbshipit-source-id: 00eb4c7091b8c7b3a48ec075b71e24aab122df85