Python: Add code formatting
Summary:
Implement a code format server call and atom provider. This currently attempts to call the yapf binary, and ignores the reformat request if not available or if an error occurs. Errors where code won't be reformatted include:
- Empty string result from the formatter (this is rarely desirable), in the case of syntax errors and otherwise.
- yapf binary is not available.

The server also searches for the closest .style.yapf file and passes it to yapf, since yapf can only search the pwd when given stdin contents instead of a file path.

Depends on D3361083

Reviewed By: hansonw

Differential Revision: D3361143

fbshipit-source-id: 3e1181498a6b5b1b5863e8fe613bfa4bea1acc08