Clang: add setting to enable diagnostics without Buck
Summary:
I was doing some local C++ development over the weekend, and needed this.
Usually diagnostics are disabled if you don't have Buck or a manual compile_commands.json, but for small one-off files the default flags are usually fine. Add a setting to enable diagnostics with default flags.

This is very useful for people using Nuclide for C++ classes as well ;)

Reviewed By: zertosh

Differential Revision: D3374436

fbshipit-source-id: 364866f7c9688f88a328404e813c25fa83e6d472