Fix text.ansi grammar horizontal scrolling
Summary:
Having a `display:none` content in Atom's text editor messes with its internal size measurement logic.
As a result it decides that on a horizontal resize of the containing panel to resize the editor contents as well. This effectively prevents any horizontal scrolling.

In `text.ansi` grammar we're hiding the color control sequences, which results in inability to horizontally scroll buck output if it has colored output.
This diff, instead of setting these sequences to `display:none` sets the size of their font to `1px` (0 didn't work) and color to transparent.

Reviewed By: matthewwithanm

Differential Revision: D3362168

fbshipit-source-id: dbbb91f756113d913be4e303d634ca0b5f77952a