Fix and re-enable nuclide-buck-base test
Summary:
This test had bit rotted. Added some `.buckversion` files and fixed the expected output of a few tests since it seems that the rule name is now included in paths. Removed all the timeout extensions, except where actually needed.

To make the test faster and less prone to timeouts:
* All `await`s have their own `waitsForPromise` block.
* Moved around a few tests so they could share the same project instance. Starting up buck is really expensive.

Reviewed By: hansonw

Differential Revision: D3363429

fbshipit-source-id: eb6e48ab28b602ef46c623a83a6fc114c918a587