Prevent Flow server from getting restarted during disposal
Summary:
A race condition would arise when the root container was cleared while a request
was in flight, and then the request would cause a new Flow server to be started.
Formalize disposal of the FlowRootContainer and prevent a disposed container
from spawning a new FlowRoot.

Note that `dispose` in the FlowService itself must clean up resources but still
allow future requests to be serviced. Both the rpc framework and the specifics
of the Flow package mean that this must currently be the case.

Reviewed By: peterhal

Differential Revision: D3380232

fbshipit-source-id: 7ba659c8322663ffb951eea286f312286a2bbd40