Kill timed out js tests at the process level
Summary: Passing the `--timeout` flag to Atom uses a `setTimeout` to kill long running tests. This works for most cases, but not for when Atom freezes. This diff uses a timer in the python runner to kill the process. It applies to all tests types. The 2 minute limit is more than generous for now.

Reviewed By: johnislarry

Differential Revision: D3388818

fbshipit-source-id: d9428aefe5875ff7eeb6c67a2fe3a0c181d190c4