Add eslint-plugin-nuclide-internal/func-params-comma-dangle rule
Summary:
This rule enforces trailing commas in multi-line function parameter lists, and forbids them in single-line function parameter lists. It also has an autofix. Remember that this is still an ecmascript proposal and not supported in Node.

There's a babel-eslint bug where arrow function params with type annotations do not get their location attributed correctly - so those don't work for this rule.

Reviewed By: nmote

Differential Revision: D3384403

fbshipit-source-id: 8ba25a0c62f228dd6009bd6670d807e54728084b