Prevent closing of modal when dragging out
Summary:
Minor annoyance - often, when attempting to select text within a modal's input field, the mouseup will occur outside of the modal, triggering a click event on the document. This closes the modal when the user is just trying to select text. Change the listeners that perform close to listen on mousedown instead of click so drags originating from within a modal won't close it. Affected modals:

- Quick Open
- File Tree Dialog

Reviewed By: shushz

Differential Revision: D3382585

fbshipit-source-id: 0c24650361c05ec5923a9e3631588afa660d3979