Console window API changes
Summary: This is the proposed api change to the console so that it can support evaluation and expandable records.  Notably, executors are now given a `getProperties` function to be used for fetching properties, and a `renderValue` function to be used by the console window for rendering values.  Both of these functions will be passed into the nested value component that jgebhardt is building

Reviewed By: jgebhardt

Differential Revision: D3231741

fbshipit-source-id: 69e88eb0a12b44e56263821c19d9611fe6ac3e49