Support opening and saving files that don't yet exist over a remote connection
Summary:
Currently, there is no use case for opening a file over a remote connection that doesn't yet exist (using `atom.workspace.open`). As such, the remote URI isn't resolved correctly, and Atom will try to open/save the `nuclide://` URI.

A buffer is successfully created to handle the URI, but the contents at the URI are attempted to be loaded and an error occurs since the file obviously doesn't exist. The cleanup involves removing the buffer and throwing an error.

This fix keeps track of the buffer for the special `ENOENT` case where the path does not yet exist on the remote server, and continues to use that buffer to deal with save events (this buffer is of a special `NuclideTextBuffer` type that knows how to deal with `nuclide://` file URIs).

NOTE: An error will be thrown by one of the Arc analytics modules the very first time you try to save, and I believe it's because it's trying to get the arcconfig for a file that doesn't yet exist. On reopening/resaving the file (once created), the error doesn't continue, so we should be okay.

Reviewed By: jgebhardt

Differential Revision: D3389455

fbshipit-source-id: bf6984bd33386e017357fec403990ce45575dd44