Change NUX flow to register via NuxManager
Summary:
Force registration of NUXes to occur through the NuxManager. The NuxManager adds the NUX to the Store, which will determine if there are duplicates. If not, the store emits an event that is read by the Manager, and the Manager adds the NUX to its list of NUXes to show.

This is done so that a Disposable can be returned by the Manager which handles removing the NUX from the pending list(s). This was not previously possible, as the NuxStore was not aware of the NUXes within the NuxManager (only the state of previously registered NUXes).

Reviewed By: jgebhardt

Differential Revision: D3369636

fbshipit-source-id: 28ec79dc65ac23e25039ec6c27d5d8edd5f247ca