Created NUX example package
Summary:
Created a sample package to demonstrate how NUXes can be integrated into packages. Requires the consumption of the `registerNux` service provided by the `nuclide-nux` package.

Creates and adds a toolbar icon, and then adds a NUX to it (for simplicity, the the same NUX is repeated twice within the tour. This can [and most likely will] be modified in a future diff).

Reviewed By: jgebhardt

Differential Revision: D3359735

fbshipit-source-id: f1f9fe51a433f6f0562313fc522e44220a464bbe