Console integration with debugger
Summary:
This hooks up the debugger to the console.  It introduces two new action pairs:

1. Add/remove console register function.  This stores/removes the function that registers an executor with the
console, allowing the debugger to call this function if it wants to provide a new executor for the console.
This happens when the debugger package is activated/deactivated.

2. Register/unregister executor.  This registers/disposes an executor provided by the debugger, which causes
the prompt in the console to appear/disappear.  This happens every time the debugger is started/stopped.

Reviewed By: jgebhardt

Differential Revision: D3360637

fbshipit-source-id: 9f2bbe55241dcb95816850a6a18e9354323003a9