Fixing Working Sets buttons tooltips
Summary:
The Button from nuclide-ui is stateless. Unfortunately this means no `ref` in React. And no `ref` means that `addTooltip` utility has nothing to work on and no tooltips are shown when hovering over working set related buttons in the file-tree.
This diff wraps the buttons in a `<span>` element and sets the tooltip to the wrapper.1

Reviewed By: ssorallen

Differential Revision: D3396984

fbshipit-source-id: 78a9e34a835a95721ab18526b9a1b8216dbd8be8