Improved seeded NUX trigger to wait for the outline view to be closed before the NUX is triggered
Summary:
The NUX triggers even if the Outline view is already open. This is bad, becuase a) it ignores the fact that the user is aware of its existence (it doesn't open without the user opening it) and b) it requires clicking twice on the Outline view button -- once to close it and once to re-open it so that the next NUX can appear.

This diff only triggers the NUX if the outline view is closed. While this doesn't address a) completely, it won't blatantly launch the NUX until the outline view is closed.

Reviewed By: nmote

Differential Revision: D3408462

fbshipit-source-id: ad065f181fef979088f9c35e9112e540a4b9ccee