Mercurial Conflict Resolver [1/n]: Detect and report conflicts
Summary:
The detection heuristic used here is:
If the repo is in the middle of a rebase, while the lock file is abscent, we query for conflicts and report it.
If the repo exits from a rebase state where it was in a conflict, then, it's no longer in a conflict state.
That will help in coming stacked diff(s) to integrate a mercurial conflict resolver in Nuclide.

Reviewed By: nmote

Differential Revision: D3395138

fbshipit-source-id: fd0423763923f4e515107fcc8dc1b6bd19c7273c