Code Format: Add option to format on save (disabled by default)
Summary:
Add a nuclide config option to format on save, which is disabled by default. This will only trigger if a code format provider is available for the current grammar.

Formatting code is an async operation within Nuclide, while atom's save is synchronous and cannot be cancelled within the onWillSave callback. Because of this, the user-triggered save operation will most likely complete before code formatting finishes. To work around this, we resave after formatting is complete, and ensure that this save operation does not cause an infinite loop by retriggering our onWillSave callback's body.

Depends on D3368190

Reviewed By: mostafaeweda

Differential Revision: D3368270

fbshipit-source-id: 487e3696d43fa71140e6a0261cac4bc6f7649018