Quick-open cycles through files from all providers integration test
Summary:
This test makes sure that the quick open:

  - Cycles through all the providers when pressing the arrow key
  - Returns back to the initially selected provider after scrolling enough

In order to make the fuzzy file service actually work during integration tests, I removed the line that was removing it from the list of available features.
johnislarry said this was initially added because it was causing tests to hang, but they seem fine now.

NOTE: This diff only tests providers that are common across all types of repositories. Specific ones, like BigGrep, Hack Symbols, will be added in another diff.

Reviewed By: johnislarry

Differential Revision: D3376944

fbshipit-source-id: 03e38181811b0bfe4ba2583b799b732697ec36b4