Introduce new package: nuclide-move-item-to-available-pane
Summary:
This package introduces commands to move panes:

```
nuclide-move-item-to-available-pane:right
nuclide-move-item-to-available-pane:left
nuclide-move-item-to-available-pane:up
nuclide-move-item-to-available-pane:down
```

I often use `cmd-k right` early on in a session to create a two column view
in Atom on my laptop. After I have done so, when I use `cmd-k right` or
`cmd-k left`, I no longer want to create a new column but move the active
pane to the opposite column [in the direction that I specified].

This revision introduces new commands to make that possible.
When `cmd-k <arrow>` is pressed, `nuclide-move-item-to-available-pane` now checks
to see if there is an existing pane in the direction of `<arrow>` (relative to the active item).
If there is, the item is moved to that pane and both the pane and item are activated.

If there is no such pane, then the existing behavior takes over: a split occurs,
creating a new pane, and the active item is moved to it.

Note that the appropriate `activationCommands` are specified in the `package.json`
such that this package is loaded lazily. It should not have any impact on startup time.

Personally, I have added the following to my `~/.atom/keymap.cson` to override the
default behavior for these keybindings in Atom on OS X:

```
'atom-text-editor':
  'cmd-k right': 'nuclide-move-item-to-available-pane:right'
  'cmd-k left': 'nuclide-move-item-to-available-pane:left'
  'cmd-k up': 'nuclide-move-item-to-available-pane:up'
  'cmd-k down': 'nuclide-move-item-to-available-pane:down'
```

This is not to say that we should make this the default choice for all Nuclide users,
but it would be nice to make users aware that this is easily configurable.

Reviewed By: zertosh, bottledwalter

Differential Revision: D3046538

fbshipit-source-id: b70ac076fc18af153a890cbe6cc8694997d09341